requirements.txt
discord.py[voice]
yt_dlp
dotenv